component/http-routing: moved AdminDashboard namespace to Admin
