+++ src/Snicco/Component/http-routing/src/Routing/Admin/AdminMenu.php
@@ -2,13 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Snicco\Component\HttpRouting\Routing\AdminDashboard;
+namespace Snicco\Component\HttpRouting\Routing\Admin;
 
 use IteratorAggregate;
+use Traversable;
 
-/**
- * @api
- */
 interface AdminMenu extends IteratorAggregate
 {
 
@@ -16,5 +14,10 @@ interface AdminMenu extends IteratorAggregate
      * @return AdminMenuItem[]
      */
     public function items(): array;
+
+    /**
+     * @return Traversable<AdminMenuItem>
+     */
+    public function getIterator(): Traversable;
 
 }
