+++ src/Snicco/Component/http-routing/tests/HttpRunnerTestCase.php
@@ -21,8 +21,8 @@ use Snicco\Component\HttpRouting\MiddlewarePipeline;
 use Snicco\Component\HttpRouting\MiddlewareStack;
 use Snicco\Component\HttpRouting\PrepareResponse;
 use Snicco\Component\HttpRouting\RouteRunner;
-use Snicco\Component\HttpRouting\Routing\AdminDashboard\AdminArea;
-use Snicco\Component\HttpRouting\Routing\AdminDashboard\WPAdminArea;
+use Snicco\Component\HttpRouting\Routing\Admin\AdminArea;
+use Snicco\Component\HttpRouting\Routing\Admin\WPAdminArea;
 use Snicco\Component\HttpRouting\Routing\Condition\RouteConditionFactory;
 use Snicco\Component\HttpRouting\Routing\Controller\FallBackController;
 use Snicco\Component\HttpRouting\Routing\Controller\RedirectController;
